backend/requirements.txt
flask
flask-cors
langchain
langchain-community
langchain-openai
langchainhub
chromadb
python-dotenv
PyPDF2
python-docx
pdfminer
rapidocr-onnxruntime
PyMuPDF
bs4